Duck Waterer 2.3
Initially used for duck water buckets that are pivoted to rotate and fill one of the 2 at a time.

This will allow a person to set up the variables for timing utilizing a switch, motor, and solenoid to dump a bucket and refill it with water automatically. Initially used for duck water buckets that are pivoted to rotate and fill one of the 2 at a time.